refactor: centralize condition evaluation
